git-svn-id: file:///home/rhornig/tmp/svn/inet-framework/trunk@1337 539eec35-cc16-0410-b457-356aab9780e1

+++ Base/ByteArrayMessage.cc
@@ -26,14 +26,14 @@ void ByteArrayMessage::setDataFromBuffer(const void *ptr, int length)
 
 void ByteArrayMessage::copyDataToBuffer(void *ptr, int length)
 {
-    ASSERT(length <= data_arraysize);
+    ASSERT((uint)length <= data_arraysize);
 
     memcpy(ptr, data_var, length);
 }
 
 void ByteArrayMessage::removePrefix(int length)
 {
-    ASSERT(data_arraysize > length);
+    ASSERT(data_arraysize > (uint)length);
     ASSERT(length > 0);
 
     int nlength = data_arraysize - length;
